Moved retry function into ReductionRunUtils
