+++ WebApp/ISIS/autoreduce_webapp/reduction_viewer/utils.py
@@ -1,10 +1,11 @@
-import logging, os, sys
+import logging, os, sys, datetime
 sys.path.append(os.path.join("../", os.path.dirname(os.path.dirname(__file__))))
 os.environ["DJANGO_SETTINGS_MODULE"] = "autoreduce_webapp.settings"
-from autoreduce_webapp.settings import LOG_FILE, LOG_LEVEL, BASE_DIR
 logger = logging.getLogger(__name__)
-from django.db import models
-from reduction_viewer.models import Instrument, Status
+from django.utils import timezone
+from reduction_viewer.models import Instrument, Status, ReductionRun, DataLocation
+from reduction_variables.models import RunVariable
+from reduction_variables.utils import InstrumentVariablesUtils
 
 class StatusUtils(object):
     def _get_status(self, status_value):
@@ -42,3 +43,60 @@ class InstrumentUtils(object):
             instrument.save()
             logger.warn("%s instrument was not found, created it." % instrument_name)
         return instrument
+        
+class ReductionRunUtils(object):
+    # def cancelRun(self, reductionRun):
+        # undefined
+
+    def createRetryRun(self, reductionRun, scripts=None, variables=None, delay=0):
+        """
+        Create a run ready for re-running based on the run provided. If variables are provided, copy them and associate them with the new one, otherwise generate variables based on the previous run. If ScriptFile objects are supplied, use them, otherwise use the previous run's.
+        """
+        # find the previous run version, so we don't create a duplicate
+        last_version = -1
+        for run in ReductionRun.objects.filter(experiment=reductionRun.experiment, run_number=reductionRun.run_number):
+            last_version = max(last_version, run.run_version)
+            
+        try:
+            # create the run object and save it
+            new_job = ReductionRun(
+                instrument = reductionRun.instrument,
+                run_number = reductionRun.run_number,
+                run_name = "",
+                run_version = last_version+1,
+                experiment = reductionRun.experiment,
+                #started_by=request.user.username, # commented out for the test server only
+                status = StatusUtils().get_queued()
+                )
+            new_job.save()
+            
+            reductionRun.retry_run = new_job
+            reductionRun.retry_when = timezone.now().replace(microsecond=0) + datetime.timedelta(seconds=delay if delay else 0)
+            reductionRun.save()
+            
+            # copy the previous data locations
+            for data_location in reductionRun.data_location.all():
+                new_data_location = DataLocation(file_path=data_location.file_path, reduction_run=new_job)
+                new_data_location.save()
+                new_job.data_location.add(new_data_location)
+                
+            if not variables: # provide variables if they aren't already
+                variables = InstrumentVariablesUtils().get_variables_for_run(new_job)
+            for var in variables:
+                new_var = RunVariable(name=var.name, value=var.value, type=var.type, is_advanced=var.is_advanced, help_text=var.help_text) # copy variable
+                new_var.reduction_run = new_job # associate it with the new run
+                new_job.run_variables.add(new_var)
+                
+                # add scripts based on whether some were supplied
+                if not scripts:
+                    scripts = var.scripts.all()
+                for script in scripts:
+                    new_var.scripts.add(script)
+                    
+                new_var.save()
+                    
+            return new_job
+            
+        except:
+            new_job.delete()
+            raise
